Trukker ut kode til egen fil
